+++ src/data/database/remote/RatingScoreTable.kt
@@ -8,7 +8,7 @@ object RatingScoreTable : Table(name = "rating_score") {
     val ratingScoreId = varchar(name = "rating_score_id", length = 50)
     val userId = varchar(name = "user_id", length = 50).references(UserTable.userId)
     val foodId = integer(name = "food_id").references(FoodTable.foodId)
-    val ratingScore = integer(name = "rating_score")
+    val ratingScore = float(name = "rating_score")
     val created = datetime(name = "created")
     val updated = datetime(name = "updated").nullable()
