+++ src/route/models/entities/RatingScoreEntity.kt
@@ -4,7 +4,7 @@ data class RatingScoreEntity(
     val ratingScoreId: String,
     val userId: String,
     val foodId: Int,
-    val ratingScore: Int,
+    val ratingScore: Float,
     val isBackup: Int,
     val created: String,
     val updated: String?,
